requirements.txt
keras==3.10.0
numpy==2.1.3
onnxruntime==1.23.0
pandas==2.3.2
pillow==11.3.0
rembg==2.0.67
streamlit==1.49.1
tensorflow==2.19.0
mediapipe==0.10.21
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.12.0.88